LayoutTests/fast/text/font-promises-gc-expected.txt

CONSOLE MESSAGE: Font status: unloaded
CONSOLE MESSAGE: Font loaded
This test should not time out, and should say "SUCCESS".
Hello
SUCCESS
